Refactoring `search` plugin
