+++ src/plugins/search/README.md
@@ -1,0 +1,15 @@
+Search plugin. it is used for custom search in text
+![search](https://user-images.githubusercontent.com/794318/34545433-cd0a9220-f10e-11e7-8d26-7e22f66e266d.gif)
+
+Disable plugin:
+```js
+Jodit.make('#editor', {
+	useSearch: false
+});
+```
+or
+```js
+Jodit.make('#editor', {
+	disablePlugins: 'search'
+});
+```
