+++ src/plugins/search/helpers/index.ts
@@ -1,0 +1,13 @@
+/*!
+ * Jodit Editor (https://xdsoft.net/jodit/)
+ * Released under MIT see LICENSE.txt in the project root for license information.
+ * Copyright (c) 2013-2022 Valeriy Chupurnov. All rights reserved. https://xdsoft.net
+ */
+
+/**
+ * @module plugins/search
+ */
+
+export * from './find-some-part-of-string';
+export * from './get-some-part-of-string-index';
+export * from './template';
