+++ src/plugins/search/search.less
@@ -22,14 +22,8 @@
 	width: 0;
 	height: 0;
 
-	visibility: hidden;
-
 	&_sticky {
 		position: fixed;
-	}
-
-	&.jodit-search_active {
-		visibility: visible;
 	}
 
 	.jodit-search__box {
